undo some changes
