+++ starsky/starsky.foundation.platform/Models/AppSettings.cs
@@ -66,9 +66,9 @@ namespace starsky.foundation.platform.Models
 
         public string BaseDirectoryProject => AppDomain.CurrentDomain.BaseDirectory
 		    .Replace("starskyadmincli", "starsky")
-		    .Replace("starskysynccli", "starsky")
+		    .Replace("starskysynchronizecli", "starsky")
 		    .Replace("starskythumbnailcli", "starsky")
-		    .Replace("starskysynclegacycli", "starsky")
+		    .Replace("starskysynccli", "starsky")
 		    .Replace("starsky.foundation.database", "starsky")
 		    .Replace("starskyImporterNetFrameworkCli", "starsky")
 		    .Replace("netframework-msbuild", "starsky")
