+++ starsky/starskysynccli/starskysynccli.csproj
@@ -1,17 +1,21 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <!-- SonarQube needs this -->
-        <ProjectGuid>{7e1136a7-cc43-49d2-91d3-48e557f0fb66}</ProjectGuid>
-        <RuntimeFrameworkVersion Condition=" '$(TargetFramework)' == 'netcoreapp3.1' ">3.1.9</RuntimeFrameworkVersion>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\starsky.foundation.database\starsky.foundation.database.csproj" />
-      <ProjectReference Include="..\starsky.foundation.platform\starsky.foundation.platform.csproj" />
-      <ProjectReference Include="..\starsky.foundation.sync\starsky.foundation.sync.csproj" />
-    </ItemGroup>
-
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <!-- SonarQube needs this -->
+    <ProjectGuid>{47d7c0fd-cee7-494c-85c2-df63234e0e5b}</ProjectGuid>
+    <RuntimeFrameworkVersion Condition=" '$(TargetFramework)' == 'netcoreapp3.1' ">3.1.9</RuntimeFrameworkVersion>
+    <DebugType>Full</DebugType>
+    <Version>0.4.0</Version>
+    <RootNamespace>starskysynccli</RootNamespace>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'" />
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'" />
+  <ItemGroup>
+    <ProjectReference Include="..\starsky.foundation.database\starsky.foundation.database.csproj" />
+    <ProjectReference Include="..\starsky.foundation.platform\starsky.foundation.platform.csproj" />
+    <ProjectReference Include="..\starsky.foundation.storage\starsky.foundation.storage.csproj" />
+    <ProjectReference Include="..\starsky.foundation.thumbnailgeneration\starsky.foundation.thumbnailgeneration.csproj" />
+    <ProjectReference Include="..\starskycore\starskycore.csproj" />
+  </ItemGroup>
 </Project>
