--- starsky/starskysynclegacycli/starskysynclegacycli.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <!-- SonarQube needs this -->
-    <ProjectGuid>{47d7c0fd-cee7-494c-85c2-df63234e0e5b}</ProjectGuid>
-    <RuntimeFrameworkVersion Condition=" '$(TargetFramework)' == 'netcoreapp3.1' ">3.1.9</RuntimeFrameworkVersion>
-    <DebugType>Full</DebugType>
-    <Version>0.4.0</Version>
-    <RootNamespace>starskysynccli</RootNamespace>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'" />
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'" />
-  <ItemGroup>
-    <ProjectReference Include="..\starsky.foundation.database\starsky.foundation.database.csproj" />
-    <ProjectReference Include="..\starsky.foundation.platform\starsky.foundation.platform.csproj" />
-    <ProjectReference Include="..\starsky.foundation.storage\starsky.foundation.storage.csproj" />
-    <ProjectReference Include="..\starsky.foundation.thumbnailgeneration\starsky.foundation.thumbnailgeneration.csproj" />
-    <ProjectReference Include="..\starskycore\starskycore.csproj" />
-  </ItemGroup>
-</Project>
